Split the test out
